backend/requirements.txt
Flask
psycopg2-binary
qrcode
reportlab
Flask-Cors